The amount in the grey background is /bin/zsh but the field is filled in as .
This creates a disconnection for me.

I think the grey background is /bin/zsh because of the due date. I think we should keep the amounts tied together despite the passing of the due date because sometimes the bill isn’t paid but it should still be visible.

With this change, the Bill Summary should then tie out. Right now the Bill Summary doesn’t tie out because of this discrepancy.